# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/dev-login": {
      "post": {
        "operationId": "AuthController_devLogin",
        "summary": "DEV ONLY: issue a signed JWT for a demo user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/citizen/otp": {
      "post": {
        "operationId": "AuthController_requestOtp",
        "summary": "Citizen login step 1: request an OTP",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/citizen/verify": {
      "post": {
        "operationId": "AuthController_verifyOtp",
        "summary": "Citizen login step 2: verify the OTP and receive a token",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpVerifyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "summary": "The authenticated caller",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/audit": {
      "get": {
        "operationId": "AuditController_list",
        "summary": "Audit entries, newest first",
        "parameters": [
          {
            "name": "entityType",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entityId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "highlighted",
            "required": false,
            "in": "query",
            "description": "true = overrides and other highlighted entries only",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "audit"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/audit/verify": {
      "get": {
        "operationId": "AuditController_verify",
        "summary": "Recompute every hash in the chain and report the first break, if any",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "audit"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/compensation": {
      "get": {
        "operationId": "CompensationController_list",
        "summary": "Compensation lines with payment references and totals by status",
        "parameters": [
          {
            "name": "projectId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "parcelId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "ASSESSED",
                "APPROVED",
                "INITIATED",
                "PAID",
                "FAILED",
                "DISPUTED",
                "ON_HOLD"
              ],
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "compensation"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/compensation/{id}/pay": {
      "post": {
        "operationId": "CompensationController_pay",
        "summary": "Disburse an approved compensation through the payment gateway (synthetic PFMS in dev)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "compensation"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/documents": {
      "get": {
        "operationId": "DocumentsController_list",
        "parameters": [
          {
            "name": "parcelId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "kind",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "GAZETTE_NOTIFICATION",
                "AWARD_COPY",
                "KHASRA_EXTRACT",
                "SIA_REPORT",
                "GRAM_SABHA_CONSENT",
                "FRA_SETTLEMENT_CERTIFICATE",
                "FOREST_CLEARANCE",
                "CRZ_CLEARANCE",
                "WILDLIFE_CLEARANCE",
                "POSSESSION_CERTIFICATE",
                "PAYMENT_ADVICE",
                "COURT_ORDER",
                "FIELD_PHOTO",
                "OTHER"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "documents"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "DocumentsController_upload",
        "summary": "Upload a document (PDF/PNG/JPEG/WebP, ≤ 10 MB). Stored with its SHA-256.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "kind": {
                    "type": "string"
                  },
                  "title": {
                    "type": "string"
                  },
                  "parcelId": {
                    "type": "string"
                  },
                  "projectId": {
                    "type": "string"
                  },
                  "referenceNo": {
                    "type": "string"
                  },
                  "issuedOn": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "documents"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/documents/{id}/download": {
      "get": {
        "operationId": "DocumentsController_download",
        "summary": "Download a document (integrity-checked against its stored SHA-256)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "documents"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "R-NLAM API",
    "description": "Real-Time National Land Acquisition & Management System. All routes need a Bearer token unless marked public. In AUTH_MODE=dev, get one from POST /api/auth/dev-login. Money is integer paise; dates are UTC.",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "DevLoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "collector.nagpur@demo.rnlam.in"
          }
        },
        "required": [
          "email"
        ]
      },
      "OtpRequestDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "9800000001",
            "description": "10-digit Indian mobile number"
          }
        },
        "required": [
          "phone"
        ]
      },
      "OtpVerifyDto": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "example": "123456"
          }
        },
        "required": [
          "requestId",
          "code"
        ]
      },
...
    }
  }
}